fix: moved StringWithValue
